requirements.txt
click
pyopenssl
pydantic